% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zr.R
\name{zr_array}
\alias{zr_array}
\title{Open an existing Zarr array}
\usage{
zr_array(store, path)
}
\arguments{
\item{store}{A ZrStore (from \code{\link[=zr_store]{zr_store()}}).}

\item{path}{Character, path within the store (e.g. \code{"/temperature"}).}
}
\value{
A ZrArray object (external pointer).
}
\description{
Open an existing Zarr array
}
