% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zr.R
\name{zr_fill_value}
\alias{zr_fill_value}
\title{Array fill value}
\usage{
zr_fill_value(arr)
}
\arguments{
\item{arr}{A ZrArray.}
}
\value{
Character representation of the fill value.
}
\description{
Array fill value
}
