% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zr.R
\name{zr_group}
\alias{zr_group}
\title{Open or create a Zarr group}
\usage{
zr_group(store, path = "/")
}
\arguments{
\item{store}{A ZrStore.}

\item{path}{Character, group path within the store.}
}
\value{
A ZrGroup object (external pointer).
}
\description{
Open or create a Zarr group
}
