% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zr.R
\name{zr_nodes}
\alias{zr_nodes}
\title{List nodes in a Zarr store}
\usage{
zr_nodes(store, path = "/")
}
\arguments{
\item{store}{A ZrStore.}

\item{path}{Character, root path to list from.}
}
\value{
A data.frame with columns \code{path} and \code{node_type}.
}
\description{
List nodes in a Zarr store
}
