% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zr.R
\name{zr_read_chunk}
\alias{zr_read_chunk}
\title{Read a single chunk}
\usage{
zr_read_chunk(arr, chunk_index)
}
\arguments{
\item{arr}{A ZrArray.}

\item{chunk_index}{Numeric vector of chunk indices (0-based).}
}
\value{
Numeric or integer vector (flat, no dimensions set).
}
\description{
Read a single chunk
}
