% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zr.R
\name{zr_store}
\alias{zr_store}
\title{Open a Zarr filesystem store}
\usage{
zr_store(path)
}
\arguments{
\item{path}{Character, path to a zarr hierarchy directory.}
}
\value{
A ZrStore object (external pointer).
}
\description{
Open a Zarr filesystem store
}
