% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zr.R
\name{zr_create_group}
\alias{zr_create_group}
\title{Create a Zarr group and store its metadata}
\usage{
zr_create_group(store, path = "/")
}
\arguments{
\item{store}{A ZrStore.}

\item{path}{Character, group path within the store.}
}
\value{
A ZrGroup object (external pointer), invisibly.
}
\description{
Create a Zarr group and store its metadata
}
