% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zr.R
\name{zr_erase_chunk}
\alias{zr_erase_chunk}
\title{Erase a chunk}
\usage{
zr_erase_chunk(arr, chunk_index)
}
\arguments{
\item{arr}{A ZrArray.}

\item{chunk_index}{Numeric vector of chunk indices (0-based).}
}
\description{
Erase a chunk
}
