% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zr.R
\name{zr_write_chunk}
\alias{zr_write_chunk}
\title{Write data to a chunk}
\usage{
zr_write_chunk(arr, data, chunk_index)
}
\arguments{
\item{arr}{A ZrArray.}

\item{data}{Numeric or integer vector.}

\item{chunk_index}{Numeric vector of chunk indices (0-based).}
}
\description{
Write data to a chunk
}
